Added basic sockets and weapon
Added in a locked chest that can only be opened after a ball has been placed into a socket on at cylinder.

Also imported in a M4 model gun and gave that a socket for a magazine.

Added three scripts:
one to ignore collisions when socketing for the gun
Another to only allow an item of a certain tag to enter the socket (only ammo tagged items for the mag socket)
And lastly one to replenish a socket with an prefabbed item, in this case I made an ammo spawner