Added UI and semi functional Lever
Added in UI elements that allow for button presses and and slider bars. The UI's are interactable and can be moved via the capsuale at the bottom of them. They also take advantage of lazy follow to always be facing the player.
There is now an interactable button that has been placed within the scene that was provided by unity as well as a custom lever built with scripts that allows for events to be triggered based on if it is on or off, there are still issues with the lever only moving down to one side and also not moving in the direction of the base plate it sits on.