Fixed Lever not working + added events
Fixed the previous issue with the lever not working, it is now fully functional with the base plate also now facing the correct direction in accordance to the levers movement. Added in another locked chest next to the lever that will start with the lid rigidbody locked, activating the lever will unlock this then deactivating the lever will lock it again, creating a toggle.