Added new test scene with throwing game testing
Added a new scene into the project and took over some of the items used in the tutorials to make a basic throwing game where the player can throw a ball at pins and knock them off the stands. Currently works but ball need large amounts of tweaking for it to have a nice throw.