More Ball game functionality + other
Small folder restructure to be more organised.
The ball game now is surrounded by a box and enclosed. There is now a collider at the bottom of the pillars that checks for pins and balls to clean them up (pins de-activated, balls destroyed)
There is also a lever at the side of the game that can toggle if the pins move side to side, these move between two points.
Added a script for the collection box and another for the moving pins.