new UI and created score manager
I have added the base script for score manager but not done any of the functionality in it yet. I added two new UI into the scene, one telling the player to flip the switch for hard mode and another that will site at the base of the game and will represent the score when the score manager is functional.