Implemented the basics for game manager
The game manager will communicate to other aspects of the game telling it how to run. For example right now it speaks to the score manager, only enabling it when the game is running and resetting the score to 0 when the game starts up. The manager also currently has the ability to activate the pins for the game when enabled and has the basics for a coroutine that controls how long the game is active for. This does need more testing.