Score and hardmode changes
Hard mode is now a bool within the game manager that is toggled with the lever, it also toggles it within the score manager through the game manager.
All score handling is fully done in the score manager now, including the score values for pins and their hard mode counterpart, this was previously held in the pin collection script.
Now when a pin falls into the zone it will call to the score manager to add a pin worth to the score, then the score manager will add a score that varies depending on if hard mode is enabled.