Working Balloon spawner
The balloon spawner and balloons are now functional. There is a pool of balloons that the spawner grabs from every few seconds, it then spawns them at one of several random spawn points. The balloons will travel to a certain destination and respawn when reached. Later when this is reached it will take health.