Built firing system not sure if working need testing
Built a firing system using the XR interaction toolkit to see when the player is holding the gun and pressing down on one of the triggers. The gun will then shoot a bullet if not on cooldown that will travel in the direction the gun is facing until hitting something or despawning.
there is now a bullet pool and the weapon and balloon systems are now organised better in the editor